split CdnMigration out from Migration
This is prep for adding AlbMigration so we can start migrating ALB
instances
